Add files via upload
The Grade Calculator is a Java-based application designed to help users easily compute their academic performance. Users can input the number of subjects along with their names and the corresponding marks obtained. The application validates the marks to ensure they fall within a realistic range of 0 to 100. It calculates the total marks, average percentage, and assigns a grade based on predefined criteria. Additionally, the program provides insights into the highest and lowest marks achieved, offering a comprehensive overview of the user's performance. This project enhances my understanding of Java programming, user input handling, and basic algorithms.